requirements.txt
flask
flask-cors
python-dotenv
pymongo
google-generativeai
PyYAML
Pillow
ensure
python-box
joblib
